refactor: reorganize package entry-point and layout.
