--- bin/snakevision
@@ -1,1 +1,0 @@
-../src/snakevision.py
