+++ snakevision/cli.py
@@ -1,84 +1,37 @@
-#!/usr/bin/env python3
+"""Command-line argument parsing utilities for snakevision.
 
+This module defines the command-line interface used by the snakevision
+package, including argument parsing, validation of user-provided style
+overrides, and generation of formatted help text describing supported
+options and customizable rendering attributes.
+"""
 # Python standard library
-from __future__ import print_function
-from abc import ABC, abstractmethod
 from dataclasses import fields, MISSING
-import argparse, sys, os, textwrap
-
+import argparse
+import os
+import sys
+import textwrap
 # 3rd party package from pypi,
 # pip install if missing
-import networkx as nx
-import dagviz
-from dagviz.style.metro import svg_renderer, StyleConfig
-
-# snakevision metadata
-__version__ = "1.0.0"
-__authors__ = "Skyler Kuhn"
-__home__ = os.path.dirname(os.path.abspath(__file__))
-_name = os.path.basename(sys.argv[0]).replace(".py", "")
-_description = "create an awesome snakemake dag!"
-_min_python = (3, 7)
-
-
-class Colors:
-    """Class encoding for ANSI escape sequeces for styling terminal text.
-    Any string that is formatting with these styles must be terminated with
-    the escape sequence, i.e. `Colors.end`.
-    """
-
-    # Escape sequence
-    end = "\33[0m"
-    # Formatting options
-    bold = "\33[1m"
-    italic = "\33[3m"
-    url = "\33[4m"
-    blink = "\33[5m"
-    higlighted = "\33[7m"
-    # Text Colors
-    black = "\33[30m"
-    red = "\33[31m"
-    green = "\33[32m"
-    yellow = "\33[33m"
-    blue = "\33[34m"
-    pink = "\33[35m"
-    cyan = "\33[96m"
-    white = "\33[37m"
-    # Background fill colors
-    bg_black = "\33[40m"
-    bg_red = "\33[41m"
-    bg_green = "\33[42m"
-    bg_yellow = "\33[43m"
-    bg_blue = "\33[44m"
-    bg_pink = "\33[45m"
-    bg_cyan = "\33[46m"
-    bg_white = "\33[47m"
+from dagviz.style.metro import StyleConfig
+# Local relative imports
+from .metadata import (
+    __version__,
+    pkg_name,
+    pkg_description,
+    pkg_min_python
+)
+from .snakevision import SnakeVision
+from .utils import (
+    Colors,
+    fatal,
+    flatten,
+    log
+)
 
 
-def err(*message, **kwargs):
-    """Prints any provided args to standard error.
-    kwargs can be provided to modify print functions
-    behavior.
-    @param message <any>:
-        Values printed to standard error
-    @params kwargs <print()>
-        Key words to modify print function behavior
-    """
-    print(*message, file=sys.stderr, **kwargs)
-
-
-def fatal(*message, **kwargs):
-    """Prints any provided args to standard error
-    and exits with an exit code of 1.
-    @param message <any>:
-        Values printed to standard error
-    @params kwargs <print()>
-        Key words to modify print function behavior
-    """
-    err(*message, **kwargs)
-    sys.exit(1)
-
-
+# Helper functions for parsing command-line arguments and building
+# custom help messages for the set of customizable style attributes.
 def parse_style_attributes(styles: str):
     """Parses custom style attributes to override the default
     styling of dagviz. For more information, please reference:
@@ -111,28 +64,6 @@ def parse_style_attributes(styles: str):
     return (k, v)
 
 
-def flatten(nested_list):
-    """Flattens a nested list. This is used to flatten a nested list
-    of parsed argparse arguments where there can be a 1:M relationship
-    between a parameter and its provided values. This allows argparse
-    options with 1:M relationships to be specified as via: `-x A -x B`
-    or `-x A B` when the append action is used with nargs=+.
-    @param nested_list <list[Any]>:
-        Nested list to flatten, any encountered list elements will be
-        flattened.
-    @returns flattened_list <list[Any]>:
-        Flattened list
-    """
-    flattened_list = []
-    for v in nested_list:
-        if isinstance(v, list):
-            # Recursively flatten nested lists
-            flattened_list.extend(flatten(v))
-        else:
-            flattened_list.append(v)
-    return flattened_list
-
-
 def build_customizable_style_attributes_help_section(left_offset_amount=18):
     """Create the help section for the set of customizable style
     attributes along with their default values. This information
@@ -154,7 +85,7 @@ def build_customizable_style_attributes_help_section(left_offset_amount=18):
     return "\n".join(stylized_attrs_defaults)
 
 
-def parsed_arguments(name, description):
+def parsed_arguments(name, description, version):
     """Parses user-provided command-line arguments. Requires argparse package.
     argparse was added to standard lib in python 3.5.
     @param name <str>:
@@ -328,178 +259,20 @@ def parsed_arguments(name, description):
         "-v",
         "--version",
         action="version",
-        version="%(prog)s {}".format(__version__),
+        version="{0} {1}".format(name, version),
         help=argparse.SUPPRESS,
     )
 
     return parser.parse_args()
 
 
-class AbstractSnakeVision(ABC):
-    """An abstract base class for defining an common interface for snakevision.
-    A snakevision class contains methods for parsing rulegraphs, building a DAG,
-    and writing the resulting dag as a snakevision image.
-
-    Usage example:
-        dag = SnakeVision('snakemake_rulegraph')
-        dag.write('/path/to/output/pipeline_dag.svg')
-    """
-
-    @abstractmethod
-    def parse(self): ...
-
-    @abstractmethod
-    def build(self): ...
-
-    @abstractmethod
-    def write(self): ...
-
-
-class SnakeVision(AbstractSnakeVision):
-    """SnakeVision class contains methods for parsing snakemake rule graphs,
-    building a DAG, and writing the resulting dag as a snakevision image.
-
-    Usage example:
-        dag = SnakeVision('snakemake_rulegraph')
-        dag.write('/path/to/output/pipeline_dag.svg')
-    """
-
-    def __init__(self, input, dag=nx.DiGraph(), skip=[], verbose=False):
-        self.input = input  # Input open file handle to a snakemake rule graph
-        self.dag = dag  # networkx.DiGraph() object to model relationships
-        self.skip = skip  # Do not add these rules in figure
-        self.verbose = verbose  # Prints parsed nodes and relationships
-        self.node2label = {}  # Maps node ids to rule names
-        self.labels = []  # Keeps track of rule name order
-        self.n2n = []  # Nested list keeps track of node-to-node relationships
-        self.verbose_dag = ""  # Verbose dag debugging information
-
-        # Parse input snakemake rule graph.
-        self.parse()
-        # Builds dag from the parsed rule graph
-        self.build()
-        # Collect and display dag debugging information
-        self.debug_dag()
-        if self.verbose:
-            print(self.verbose_dag)
-
-    def __repr__(self):
-        return textwrap.dedent(
-            """
-            Class usage example:
-            dag = SnakeVision('snakemake_rulegraph.txt')
-            dag.write('/path/to/output/pipeline_dag.svg')
-            """
-        )
-
-    def __str__(self):
-        return self.verbose_dag
-
-    def debug_dag(self):
-        """Displays verbose information of the parsed rule graph."""
-        _c = Colors
-        self.verbose_dag = textwrap.dedent(
-            """
-            {3}{4}Node-to-label mappings:{5}
-            • {0}
-
-            {3}{4}Node to node relationship:{5}
-            • {1}
-
-            {3}{4}Order of encountered labels:{5}
-            • {2}
-            """.format(
-                self.node2label, self.n2n, self.labels, _c.bold, _c.url, _c.end
-            )
-        )
-
-    def parse(self):
-        """Parses a snakemake rule graph, saves parsed output in node2labels
-        labels, and n2n instance variables.
-
-        Example input rule graph:
-           0[label = "all", color = "0.34 0.6 0.85", style="rounded"];
-           1[label = "fc_lane", color = "0.35 0.6 0.85", style="rounded"];
-           2[label = "fastqc_raw", color = "0.09 0.6 0.85", style="rounded"];
-           1 -> 0
-           2 -> 0
-        """
-        for line in self.input:
-            line = line.strip()
-            # Skip over empty and un-important
-            # lines, rulegraph node ids start
-            # with a number
-            if not line:
-                continue
-            if not line[0].isdigit():
-                continue
-            # Parse nodes and edges in dag,
-            # need to map a node id to a
-            # snakemake rule name
-            if "[" in line:
-                # node definition line, i.e:
-                # 0[label = "all", color = "0.34 0.6 0.85", style="rounded"];
-                node = int(line.split("[")[0])
-                label = (
-                    line.split("=")[1]
-                    .split(",")[0]
-                    .strip()
-                    .replace('"', "")
-                    .replace("'", "")
-                )
-                if label == self.skip:
-                    # Get node ID of skip node, i.e
-                    # rule all or the final target
-                    skipnode = node
-                self.node2label[node] = label
-                self.labels.append(label)
-            elif "->" in line:
-                # node relationships or edges, i.e:
-                # 1 -> 0
-                nodes = [int(n.strip()) for n in line.split("->")]
-                lab1 = self.node2label[nodes[0]]
-                lab2 = self.node2label[nodes[1]]
-                self.n2n.append([lab1, lab2])
-
-    def build(self):
-        """Builds a DAG from a parsed snakemake rule graph."""
-        # Add nodes to the dag
-        for lab in self.labels:
-            if lab in self.skip:
-                # Do NOT include rule all
-                # in the final figure
-                continue
-            self.dag.add_node(lab)
-
-        # Add relationships or edges to dag
-        for l1, l2 in self.n2n:
-            if l1 in self.skip or l2 in self.skip:
-                # Do NOT include rule all
-                # in the final figure
-                continue
-            self.dag.add_edge(l1, l2)
-
-    def write(self, output="snakevision_dag.svg", style={}):
-        """Write DAG to output image."""
-        # Create output directory as needed
-        outdir = os.path.dirname(os.path.abspath(output))
-        if not os.path.exists(outdir):
-            # Pipeline output directory
-            # does not exist on filesystem
-            os.makedirs(outdir)
-        # Apply a custom style if available
-        custom_style = svg_renderer(StyleConfig(**style))
-        # Write snakevision dag to SVG
-        o = dagviz.render_svg(self.dag, style=custom_style)
-        # Write SVG to disk
-        with open(output, "wt") as fs:
-            fs.write(o)
-
-
 def main():
+    """Entry-point to the snakevision command-line interface."""
+    # Main function/entry point for snakevision,
+    # called when executing `python -m snakevision`.
     # Check for minimum python verison,
     # enforces running with python>=3.7
-    MINIMUM_PYTHON = _min_python
+    MINIMUM_PYTHON = pkg_min_python
     MAJOR_VERSION, MINOR_VERSION = MINIMUM_PYTHON
     if sys.version_info < MINIMUM_PYTHON:
         fatal(
@@ -509,7 +282,11 @@ def main():
         )
 
     # Parse command-line arguments
-    args = parsed_arguments(name=_name, description=_description)
+    args = parsed_arguments(
+        name=pkg_name,
+        description=pkg_description,
+        version=__version__
+    )
 
     # Flatten options witn a 1:M relationships.
     # This allows options to be specified as
@@ -519,13 +296,17 @@ def main():
     args.style = flatten(args.style)
 
     # Create a snakevision dag object
+    log("Started parsing rulegraph: {0}".format(getattr(args.input_file, "name", "")))
+    if args.skip_rules:
+        log("Skipping over the following rules in DAG visualization: {0}".format(args.skip_rules))
     dag = SnakeVision(
         input=args.input_file, skip=args.skip_rules, verbose=args.debug_dag
     )
+    log("Finished parsing rulegraph: {0}".format(getattr(args.input_file, "name", "")))
 
     # Write dag object to file
+    log("Started writing DAG to the following output file: {0}".format(args.output))
+    if args.style:
+        log("Applying additional styling options to DAG: {0}".format(args.style))
     dag.write(output=args.output, style=dict(args.style))
-
-
-if __name__ == "__main__":
-    main()
+    log("Finished writing DAG to the following output file: {0}".format(args.output))
